Refactor header and remove unused pages
